refactor: Use selectors to compute categories data
